[#500] Refactor reindexing infrastructure
We're starting to have many different indexes, so it's about time to DRY them
up.

opentrials/opentrials#500
