+++ tools/indexers/index.js
@@ -1,0 +1,6 @@
+'use strict';
+
+module.exports = {
+  trials: require('./trials'),
+  autocomplete: require('./autocomplete'),
+}
